Eliminated pyatlan/client/typedef.py
