+++ pyatlan/cache/classification_cache.py
@@ -1,7 +1,6 @@
 from typing import Optional
 
 from pyatlan.client.atlan import AtlanClient
-from pyatlan.client.typedef import TypeDefClient
 from pyatlan.model.enums import AtlanTypeCategory
 from pyatlan.model.typedef import ClassificationDef
 
@@ -16,9 +15,7 @@ class ClassificationCache:
 
     @classmethod
     def _refresh_cache(cls) -> None:
-        response = TypeDefClient(AtlanClient()).get_typedefs(
-            type=AtlanTypeCategory.CLASSIFICATION
-        )
+        response = AtlanClient().get_typedefs(type=AtlanTypeCategory.CLASSIFICATION)
         if response is not None:
             cls.cache_by_id = dict()
             cls.map_id_to_name = dict()
